More type fixes. That should be all (for now).

+++ src/NewVoronoiGrid.hpp
@@ -203,7 +203,7 @@ private:
      * NewVoronoiGridConstructionJob.
      */
     inline NewVoronoiGridConstructionJob *get_job(int_fast32_t thread_id) {
-      const size_t cellsize = _grid._cells.size();
+      const uint_fast32_t cellsize = _grid._cells.size();
       if (_current_index == cellsize) {
         return nullptr;
       }
